+++ src/components/app/app.ts
@@ -5,7 +5,7 @@ import MainView from '../view/main/main-view';
 import Router from '../router/router';
 import type { Route } from '../router/router';
 import { Pages } from '../router/pages';
-import OptionsView from '../view/main/options/options';
+import OptionsView from '../view/main/options/options-view';
 import DecisionView from '../view/main/decision/decision';
 import NotFoundView from '../view/main/not-found/not-found-view';
